+++ Changes.md
@@ -7,6 +7,9 @@ Improvements
 - OSLShader :
  - Improved loading of spline parameters with additional duplicate endpoints.
  - Added support for loading splines from RenderMan shaders.
+- RenderManShader :
+ - Added support for loading splines.
+
 
 Fixes
 -----